Settings
Settings menu creates a config file that saves some settings. It saves directly to the assets directory right now because I wasn't sure what folder it could fit in.

There are also some minor changes to the 2D HUD to make it easier to work with, like proper component names.