Different tower selection + other changes
You can now place different towers depending on the one you select in the 2D HUD. Each time a new tower is added to the game, the match selected_tower needs to be expanded to account for it.

The tower info dictionary now has accurate information.

game_manager has a _change_gold function that adds or subtracts a value to the gold count. The debug key G adds 100 gold.